refactor(openomni): inline protocol-only fixture and remove root tests/
