--- tests/fixtures/protocol-only/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@openomni-fixtures/protocol-only",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
-  "scripts": {
-    "build": "tsc --noEmit",
-    "check-types": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@openomni/protocol": "workspace:*"
-  }
-}
