--- tests/fixtures/protocol-only/src/index.ts
@@ -1,63 +1,0 @@
-import { AgentProfile, Extension, McpConfig, Skill, Tool } from "@openomni/protocol";
-
-const agent: AgentProfile.Definition = AgentProfile.Definition.parse({
-  name: "protocol-fixture-agent",
-  description: "Exercises protocol-only extension authoring.",
-  systemPrompt: "Use only protocol contracts when describing fixture behavior.",
-  tools: ["protocol_fixture.inspect"],
-  model: { provider: "fixture", id: "deterministic" },
-  budget: { maxTurns: 1, maxToolCalls: 1 },
-});
-
-const tool: Tool.Spec = Tool.Spec.parse({
-  name: "protocol_fixture.inspect",
-  description: "Returns deterministic fixture metadata.",
-  inputSchema: {
-    type: "object",
-    properties: {
-      subject: { type: "string" },
-    },
-    required: ["subject"],
-  },
-  safe: true,
-});
-
-const skill: Skill.Definition = Skill.Definition.parse({
-  id: "protocol-fixture-skill",
-  name: "Protocol Fixture Skill",
-  description: "Keeps extension fixture behavior deterministic.",
-  scope: "local",
-  layer: "execution",
-  path: "./skills/protocol-fixture/SKILL.md",
-  promptFragment: "Validate manifest contracts without runtime imports.",
-  finalChecklist: ["Manifest parses through protocol schemas."],
-});
-
-const mcpServer: McpConfig.ServerConfig = McpConfig.ServerConfig.parse({
-  name: "protocol-fixture-mcp",
-  transport: "stdio",
-  command: "protocol-fixture-mcp",
-  args: ["--deterministic"],
-  timeout: 1_000,
-  retries: 0,
-});
-
-export const manifest: Extension.Manifest = Extension.Manifest.parse({
-  id: "protocol-only-fixture",
-  name: "Protocol Only Fixture",
-  version: "1.0.0",
-  description: "A tiny extension manifest authored with protocol exports only.",
-  author: "OpenOmni Test Fixture",
-  contributes: {
-    agents: [agent],
-    tools: [tool],
-    skills: [skill],
-    mcpServers: [mcpServer],
-  },
-  compatibility: { openomni: ">=0.1.0" },
-  provenance: { manifestHash: "protocol-only-fixture-1.0.0" },
-});
-
-export function getManifest(): Extension.Manifest {
-  return manifest;
-}
